+++ examples/service_worker/public/subscribe.js
@@ -8,11 +8,11 @@
 //            the push server will use to authenticate your application server.
 //            - https://developer.mozilla.org/en-US/docs/Web/API/PushManager/subscribe
 // @returns Promise<web_sys::PushSubscription>
-// Until https://github.com/rustwasm/wasm-bindgen/pull/2288 is included in the new release of wasm_bindgen,
+// Until https://github.com/rustwasm/wasm-bindgen/pull/2288 is included in the new release of wasm-bindgen,
 // this function should be called from the seed app to subscribe to the PushManager. The applicationServerKey 
 // is the vapid key that is used for identification on the back-end server. The `userVisibleOnly` property
 // indicates that the returned push subscription will only be used for messages whose effect is made visible 
-// to the user. It must be set to `true` or the browser will reject the subscripbtion request. This holds true
+// to the user. It must be set to `true` or the browser will reject the subscription request. This holds true
 // for both chrome and firefox.
 // ------ ------
 window.subscribe = async (manager, api_key) => {
@@ -21,6 +21,6 @@ window.subscribe = async (manager, api_key) => {
         userVisibleOnly: true
     });
 
-    console.log("subscription", subscription);
+    console.log("JS subscription", subscription);
     return subscription;
 }